+++ CHANGE_LOG.md
@@ -4,6 +4,26 @@ This log contains changes made to the project. Each entry contains changed made 
 If this is a problem, use exact build numbers to track changes. As each build logs the git-hash it was created from to better understand changes made.
 
 # Versions
+## 3.#.#
+### Runtime Changes
+
+### Development Changes
+
+## 3.0.2
+### Runtime Changes
+Added: Configs for content
+Added: Settings to disable radiation system
+Added: ISidedInventory support (hopper and pipe support) for reactor cores 
+Added: Mechanic to move lowest time rod to bottom of reactor stack
+Added: Damage display to fuel rods
+Added: Display info to fuel rods
+
+Improved: handling of entity item radiation tracking
+
+Fixed: GUI localizations
+Fixed: NPE when checking if can output fluids from processing machines
+Fixed: Fuel rods being usable beyond time limit (timer going negative)
+
 ## 3.0.1
 ### Changes
 Full rewrite not based on the code from 1.6.4, 1.5.2, or Resonant Induction versions